% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detct_RstCor.R
\name{detct_RstCor}
\alias{detct_RstCor}
\title{Detect Raster Correlation}
\usage{
detct_RstCor(Stk, THvalue, returnCorTab = FALSE)
}
\arguments{
\item{Stk}{a RasterStack}

\item{THvalue}{numeric as 0.X - threshold value for correlation value of correlating layers to drop correlating layers (see details).}

\item{returnCorTab}{boolean - to return either the cleaned RasterStack (FALSE) or to return the correlation table (TRUE); default= FALSE}
}
\value{
returns a RasterStack with the layers correlating less than the treshold value.
}
\description{
detects correlation of RasterLayers in a RasterStack
}
\details{
This function is used to test a RasterStack on the correlation of the RasterLayers. All RasterLayers which have a higher correlation value than 'THvalue'
will be dropped from the RasterStack. E.g. if THvalue=0.9 all RasterLayers with correlation values >0.9 and  < -0.9 will be dropped.
}
\note{
\itemize{
\item To perform a correlation test all values are cleaned of INF and/or NA values.
\item The output RasterStack is NOT cleaned from INF and/or NA values.
}
}
\examples{
### load data to compute RGB Indices
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)
names(mspec)<- c("blue","green","red","nir")
### compute all vegetation indizes
x <-LEGION::vegInd_RGB(exp_rgb,3,2,1)
# perform Cor Test
y <- detct_RstCor(x,0.7)
names(y)
# to return the Correlation Matrix
z <- detct_RstCor(x,0.7,returnCorTab=TRUE)
}
\author{
Andreas Schönberg
}
