% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Rst.R
\name{filter_Rst}
\alias{filter_Rst}
\title{filter single band raster}
\usage{
filter_Rst(rst, fLS = "all", sizes)
}
\arguments{
\item{rst}{a single band raster layer}

\item{fLS}{comma separated character for desired filter functions. Select from
"sum","min","max","sd","mean","modal","sobel","sobel_hrzt","sobel_vert","sobel"
default = all (see details for further informations)}

\item{sizes}{numeric values for the moving window, must be odd}
}
\value{
Returns a raster stack with the selected filters
}
\description{
calculates several filters for a single raster
}
\details{
\subsection{available filter methods}{
\itemize{
\item "sum" - Sum of all cells in Moving Window
\item "min" - minimum value in Moving Window
\item "max" - maximum value in Moving Window
\item "sd"  - standard deviation
\item "modal" - modal
\item "sobel" - sobel filter in horizontal and vertical directions
\item "sobel_hrzt" - horizontal only
\item "sobel_vert" - vertical only
}
}
}
\note{
notes if needed
}
\examples{
### load data
data(exp_rst)
### compute all filter
x <- filter_Rst(exp_rst,sizes=c(3,5,7))
plot(x[[]])
### compute specific filters
flist <- c("modal","sobel_vert","mean")
y <- filter_Rst(exp_rst,fLS=flist,sizes=c(3,5,7))
}
\seealso{
\code{\link{focal}}
}
\author{
Andreas Schönberg
}
