% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Stk.R
\name{filter_Stk}
\alias{filter_Stk}
\title{wrapper for 'filter_Rst' to filter RasterStacks}
\usage{
filter_Stk(stk, fLS = "all", sizes, layernames = names(stk))
}
\arguments{
\item{fLS}{comma separated character for desired filter functions. Select from
"sum","min","max","sd","mean","modal","sobel","sobel_hrzt","sobel_vert","sobel"
default = all (see details for further informations)}

\item{sizes}{numeric values for the moving window, must be odd}

\item{layernames}{optional - comma seperated character for desired layernames}

\item{Stk}{a Raster Stack}
}
\value{
Returns a raster stack with the selected filters
}
\description{
calculates several filters for each Single Raster in a RasterStack
}
\details{
\subsection{available filter methods}{
\itemize{
\item "sum" - Sum of all cells in Moving Window
\item "min" - minimum value in Moving Window
\item "max" - maximum value in Moving Window
\item "mean"- mean of all cells in Moving Window
\item "sd"  - standard deviation
\item "modal" - modal
\item "sobel" - sobel filter in horizontal and vertical directions
\item "sobel_hrzt" - horizontal only
\item "sobel_vert" - vertical only
}
}
}
\note{
notes if needed
}
\examples{
### load data
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)

### compute all filters or every layer in stack
x <- filter_Stk(mspec,sizes=3)
x # note that the names are basic set x.1
### define layernames
ln <- c("blue","green","red","nir")
z <- filter_Stk(mspec,fLS="sum",sizes=3,layernames=ln)
names(z)
### compute specific filters
flist <- c("modal","sobel_vert","mean")
y <- filter_Stk(mspec,fLS=flist,sizes=c(3,5,7))
y
}
\seealso{
\code{\link{focal}},\code{\link{filter_Rst}}
}
\author{
Andreas Schönberg
}
