% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGION-package.R
\docType{data}
\name{lau_mspec - data}
\alias{lau_mspec - data}
\title{A multispectral image (RGB+nir) with some trees in the lautaret vally in the france alps.}
\format{
\code{"raster::raster"}
}
\description{
An RGB orthophoto merged with the Nir band of a IRC image. Bands sequence is: 1-blue,2-green,3-red,4-nir. Resolution 0.15 meter.
}
