% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGION-package.R
\docType{package}
\name{LEGION - package}
\alias{LEGION - package}
\title{LEGION - a tool for computing several artificial layers (AL)}
\description{
In general the \code{LEGION} package provides tools to compute a huge stack of artifical layers
to use in maschine learing or ordination/cluster analysis. These methods require lots of information, the more the better (unless it leads to overfitting).
With \code{LEGION} the user can first compute several indices and apply filter to generate artifical layers of a RasterLayer or a RasterStack.
Further \code{LEGION} provides a function to detect higly correlating layers (which could lead to overfitting and/or high computing time while processing).
}
\details{
LEGION
}
\note{
In general a 'brute force' use of \code{LEGION} is not recommended. E.g. applying all available filters with multiple sizes of MovingWindows on all availabe indices would lead to multiple hundreds of artifal layers.
This could take hours to days to compute (especially in the case of large RasterStacks) which makes little sense.

On the other hand  \code{LEGION} helps to reduce "overthinking" in planning which ALs should be used for a maschine learing or ordination/cluster analysis task.
A recommended way to generate a RasterStack with the required information but minimal correlation in a relatively short time is to first generate the indices and then to clean the RasterStack from correlating ALs and then successively to filter the cleaned RasterStack.
}
\author{
Andreas Schönberg
}
\keyword{package}
