% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legion-package.R
\docType{package}
\name{smooth}
\alias{smooth}
\title{legion package}
\description{
Package contains functions for multivariate time series forecasting
}
\details{
\tabular{ll}{ Package: \tab smooth\cr Type: \tab Package\cr Date: \tab
2016-01-27 - Inf\cr License: \tab GPL-2 \cr } The following functions are
included in the package:
\itemize{
\item \link[smooth]{ves} - Vector Exponential Smoothing.
\item \link[smooth]{ves} - Vector ETS-PIC model.
\item \link[smooth]{viss} - Multivariate occurrence ETS model.
}
}
\references{
\itemize{
\item de Silva A., Hyndman R.J. and Snyder, R.D. (2010). The vector
innovations structural time series framework: a simple approach to
multivariate forecasting. Statistical Modelling, 10 (4), pp.353-374
\item Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D. (2008)
Forecasting with exponential smoothing: the state space approach,
Springer-Verlag.
\item Lütkepohl, H. (2005). New Introduction to Multiple Time Series
Analysis. New introduction to Multiple Time Series Analysis. Berlin,
Heidelberg: Springer Berlin Heidelberg.
\doi{10.1007/978-3-540-27752-1}
\item Chen H., Svetunkov I., Boylan J. (2021). A New Taxonomy of Vector
Exponential Smoothing for Seasonal Time Series.
}
}
\seealso{
\code{\link[greybox]{forecast}, \link[smooth]{es}, \link[smooth]{adam}}
}
\author{
Ivan Svetunkov

Kandrika Pritularga

Maintainer: Ivan Svetunkov <ivan@svetunkov.ru>
}
\keyword{models}
\keyword{multivariate}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
