% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegInd_Nir.R
\name{vegInd_dev}
\alias{vegInd_dev}
\alias{vegInd_RGB}
\title{calculate RGB+Nir vegetation indices}
\usage{
vegInd_dev(rgb, indlist = "all")
}
\arguments{
\item{rgb}{a RasterStack or Brick with RGB bands}

\item{indlist}{comma separated character for desired Vegetation Indices to compute. Select from
"NDVI","TDVI","SR","MSR" default=all (see details for further information)}
}
\value{
Returns a raster stack with the selected Vegetation Indices
}
\description{
computes several Vegetation Indices based on RGB bands
}
\details{
\subsection{available indices}{

"NDVI"
}
}
\note{
notes if needed
}
\references{
The IDB Project (2020): Index Database (https://www.indexdatabase.de/)
}
\author{
Andreas Schönberg
}
