% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegInd_Nir.R
\name{vegInd_mspec}
\alias{vegInd_mspec}
\title{calculate RGB+Nir vegetation indices}
\usage{
vegInd_mspec(mspec, indlist = "all")
}
\arguments{
\item{mspec}{a RasterStack or Brick with multispectral RGB+Nir bands}

\item{indlist}{comma separated character for desired Vegetation Indices to compute. Select from
"NDVI","TDVI","SR","MSR" default=all (see details for further information)}
}
\value{
Returns a raster stack with the selected Vegetation Indices
}
\description{
computes several Vegetation Indices based on RGB bands
}
\details{
\subsection{available indices}{

"NDVI"
}
}
\note{
notes if needed
}
\examples{
### load data
exp_mspec <-LEGION::exp_mspec
### compute all vegetation indizes
x <-LEGION::vegInd_mspec(exp_mspec)
plot(x)
### select specific vegetation indices
vi <-c("NDVI","TDVI")
y <-LEGION::vegInd(exp_rgb,indlist=vi)
plot(y)
}
\references{
The IDB Project (2020): Index Database (https://www.indexdatabase.de/)
}
\author{
Andreas Schönberg
}
