% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegInd_RGB.R
\name{vegInd_RGB}
\alias{vegInd_RGB}
\title{calculate RGB vegetation indices}
\usage{
vegInd_RGB(rgb, red = NULL, green = NULL, blue = NULL, indlist = "all")
}
\arguments{
\item{rgb}{a RasterStack or Brick with RGB bands}

\item{red}{the band/layer number of band'red'}

\item{green}{the band/layer number of band'green'}

\item{blue}{the band/layer number of band'blue'}

\item{indlist}{comma separated character for desired Vegetation Indices to compute. Select from
"VVI","VARI","NDTI","RI","CI","BI","SI","HI","TGI","GLI","NGRDI", default=all (see details for further information)}
}
\value{
Returns a raster stack with the selected Vegetation Indices
}
\description{
computes several Vegetation Indices based on RGB bands
}
\details{
\subsection{available indices}{

"VVI"
"VARI"
"NDTI"
"RI"
"CI"
"BI"
"SI"
"HI"
"TGI"
"GLI"
"NGRDI"
}
}
\note{

}
\examples{
### load data
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)
names(mspec)<- c("blue","green","red","nir")
### compute all vegetation indizes
x <-LEGION::vegInd_RGB(mspec,3,2,1)
plot(x)
### select specific vegetation indices
vi <-c("VVI","SI","GLI")
y <-LEGION::vegInd_RGB(mspec,3,2,1,indlist=vi)
plot(y)
}
\references{
The IDB Project (2020): Index Database (https://www.indexdatabase.de/)
}
\author{
Andreas Schönberg
}
