% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detct_RstCor.R
\name{detct_RstCor}
\alias{detct_RstCor}
\title{detect Raster Correlation}
\usage{
detct_RstCor(Stk, THvalue, returnCorTab = FALSE)
}
\arguments{
\item{Stk}{a RasterStack}

\item{THvalue}{numeric in 0.X - Treshold Value for Correlation Value to drop Layers.}

\item{returnCorTab}{boolean - to return either the cleand Stack (default=FALSE) or to return the Correlation Table with TRUE}
}
\value{
Returns a raster stack with the layers correlation less than the Treshold Value.
}
\description{
detects Correlation of Rasterlayers in a RasterStack and returns a Rasterstack without the correlating Layers
}
\details{
This function is used to test a RasterStack for correlating Layers. All layers which have a greater Correlation than the THvalue
will be dropped from the Stack. E.G. if THvalue=0.9 all Layers with corvalues of >0.9 and <-0.9 will be dropped.
}
\note{
\itemize{
\item To perform a correlation Test all Values are cleaned from INF and or NA Values.
\item The output Stack is NOT cleaned from INF or NA Values.
}
}
\examples{
### load data to compute RGB Indices
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)
names(mspec)<- c("blue","green","red","nir")
### compute all vegetation indizes
x <-LEGION::vegInd_RGB(exp_rgb,3,2,1)
# perform Cor Test
y <- detct_RstCor(x,0.7)
names(y)
# to return the Correlation Matrix
z <- detct_RstCor(x,0.7,returnCorTab=TRUE)
}
\author{
Andreas Schönberg
}
