% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detct_RstHgmy.R
\name{detct_RstHmgy}
\alias{detct_RstHmgy}
\title{detect Raster homogeneity}
\usage{
detct_RstHmgy(Stk, THvalue)
}
\arguments{
\item{Stk}{a RasterStack}

\item{THvalue}{numeric  - in percent (0.x) Treshold Value for homogeneity Value to drop Layers.}
}
\value{
Returns the RasterStack without homogenious RasterLayers.
}
\description{
detects homogeneity of Rasterlayers in a RasterStack and drops RasterLayers with homogeneity higher than a Treshold Value.
}
\details{
This function is used to test a RasterStack for homogenious RasterLayers. All RasterLayers with any values which represent more or equal percentage (THvalue) of all cells are dropped.
E.g. if a RasterLayer with ncell= 100 contains 95 cell with value 1 it will be dropped if the THvalue is 0.95 or less.
}
\note{
\itemize{
\item The output Stack is NOT cleaned from INF or NA Values.
}
}
\examples{
### load data
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)
names(mspec)<- c("blue","green","red","nir")
### compute all vegetation indizes
x <-LEGION::vegInd_RGB(mspec,3,2,1)
plot(x)
### test for homogeneity with 90\% occurance of any cell value
hmgy90 <-detct_RstHmgy (x,0.9)
hmgy90
### test for homogeneity with 10\% occurance of any cell value
hmgy10 <-detct_RstHmgy (x,0.1)
hmgy10 # returns emtpy RasterStack
### test for homogeneity with 100\% occurance of any cell value
hmgy90 <-detct_RstHmgy (x,1.0)
hmgy90 # returns full RasterStack
}
\author{
Andreas Schönberg
}
