% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Rst.R
\name{filter_Rst}
\alias{filter_Rst}
\title{Filter Single-band RasterLayer}
\usage{
filter_Rst(rst, fLS = "all", sizes, NArm = TRUE)
}
\arguments{
\item{rst}{a single-band RasterLayer.}

\item{fLS}{comma-separated character combinations of the desired filter functions. Select from
"sum","min","max","sd","mean","modal","sobel","sobel_hrzt","sobel_vert".
default = all; (see details).}

\item{sizes}{numeric - a single value or combinations  for the MovingWindow, number must be odd.}

\item{NArm}{boolean - removes NA values, default= TRUE.}
}
\value{
returns a RasterStack with the desired filtered artificial layers.
}
\description{
applies several filter functions on single RasterLayer.
}
\details{
\subsection{available filter functions}{
\itemize{
\item "sum" - sum of all cells in a MovingWindow
\item "min" - minimum value of all cells in a MovingWindow
\item "max" - maximum value of all cells in a MovingWindow
\item "mean"- mean value of all cells in a MovingWindow
\item "sd"  - standard deviation of all cells in a MovingWindow
\item "modal" - most frequent value of all cells in a MovingWindow
\item "sobel" - sobel edge detection filter in horizontal and vertical directions
\item "sobel_hrzt" - sobel edge detection filter in horizontal direction only
\item "sobel_vert" - sobel edge detection filter in vertical direction only
}
}
}
\examples{
### load data
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)
names(mspec)<- c("blue","green","red","nir")
### seperate single raster layer
rst <- mspec$nir
### compute all filter
x <- filter_Rst(rst,sizes=c(3,5,7))
plot(x[[3]])
### compute specific filters
flist <- c("modal","sobel_vert","mean")
y <- filter_Rst(rst,fLS=flist,sizes=c(3,5,7))
}
\seealso{
\code{\link{focal}}
}
\author{
Andreas Schönberg
}
