% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Rst.R
\name{filter_Rst}
\alias{filter_Rst}
\title{filter_Rst}
\usage{
filter_Rst(rst, fLS, sizes)
}
\arguments{
\item{rst}{a single band raster layer}

\item{fLS}{comma separated character for desired filter functions. Select from
"sum","min","max","sd","mean","modal","sobel","sobel_hrzt","sobel_vert","sobel"
default = all (see details for further informations)}

\item{sizes}{numeric values for the moving window, must be odd}
}
\value{
Returns a raster stack with the selected filters
}
\description{
calculates several filters for a single raster
}
\details{
further description
}
\note{
notes if needed
}
\examples{
# need to lern and test this
}
\references{
Meyer, H., Reudenbach, C., Hengl, T., Katurji, M., Nauß, T. (2018): Improving performance of spatio-temporal machine learning models using forward feature selection and target-oriented validation. Environmental Modelling & Software 101: 1-9.
}
\seealso{
\code{\link{trainControl}},\code{\link{ffs}} ### link to other functions
}
\author{
Andreas Schönberg
}
