% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Rst.R
\name{filter_Rst}
\alias{filter_Rst}
\title{filter single band raster}
\usage{
filter_Rst(rst, fLS = "all", sizes)
}
\arguments{
\item{rst}{a single band raster layer}

\item{fLS}{comma separated character for desired filter functions. Select from
"sum","min","max","sd","mean","modal","sobel","sobel_hrzt","sobel_vert","sobel"
default = all (see details for further informations)}

\item{sizes}{numeric values for the moving window, must be odd}
}
\value{
Returns a raster stack with the selected filters
}
\description{
calculates several filters for a single raster
}
\details{
\subsection{Subsection within details}{
\subsection{Sub-subsection}{

... text ...Filter arguments:
\itemize{
\item "sum" - Sum of all cells in MovingWindow
\item "min" - minimum value
\item "max" - maximum vlaue
}
\enumerate{
\item sd"  - standart deviation
\item modal"- modal
3 . sobel" - sobel filter in horizontal and vertical directions
| "sobel_hrzt"| - horizontal only
| "sobel_vert"| - vertical only
}
}

}
}
\note{
notes if needed
}
\examples{
### load data
data(exp_rst)
### compute all filter
x <- filter_Rst(exp_rst,sizes=c(3,5,7))
plot(x[[]])
### compute specific filters
flist <- c("modal","sobel_vert","mean")
y <- filter_Rst(exp_rst,fLS=flist,sizes=c(3,5,7))
}
\seealso{
\code{\link{focal}}
}
\author{
Andreas Schönberg
}
