% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Stk.R
\name{filter_Stk}
\alias{filter_Stk}
\title{wrapper for 'filter_Rst' to filter RasterStacks}
\usage{
filter_Stk(stk, fLS = "all", sizes, layernames = names(stk))
}
\arguments{
\item{fLS}{comma separated character for desired filter functions. Select from
"sum","min","max","sd","mean","modal","sobel","sobel_hrzt","sobel_vert","sobel"
default = all (see details for further informations)}

\item{sizes}{numeric values for the moving window, must be odd}

\item{layernames}{optional - comma seperated character for desired layernames}

\item{Stk}{a Raster Stack}
}
\value{
Returns a raster stack with the selected filters
}
\description{
calculates several filters for each Single Raster in a RasterStack
}
\details{
further description
}
\note{
notes if needed
}
\examples{
### load data
data(exp_rgb)
### compute all filter
x <- filter_Stk(exp_rgb,sizes=3)
x
### compute specific filters
flist <- c("modal","sobel_vert","mean")
y <- filter_Stk(exp_rst,fLS=flist,sizes=c(3,5,7))
y
### define layernames
ln <-c("red","green","blue")
z <- filter_Stk(exp_rgb,fLS="sum",sizes=3,layernames=ln)
names(z)
}
\seealso{
\code{\link{focal}},\code{\link{filter_Rst}}
}
\author{
Andreas Schönberg
}
