% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Stk.R
\name{filter_Stk}
\alias{filter_Stk}
\title{Wrapper for 'filter_Rst' to filter RasterStacks}
\usage{
filter_Stk(stk, fLS = "all", sizes, layernames = names(stk))
}
\arguments{
\item{fLS}{comma-separated character combinations of the desired filter functions. Select from
"sum","min","max","sd","mean","modal","sobel","sobel_hrzt","sobel_vert".
default = all; (see details).}

\item{sizes}{numeric - a single value or combinations for the MovingWindow, number must be odd.}

\item{layernames}{optional - comma-seperated character combinations of the desired layernames.}

\item{Stk}{a RasterStack.}
}
\value{
returns a RasterStack with the desired filtered artificial layers.
}
\description{
applies several filter functions to each RasterLayer in a RasterStack.
}
\details{
\subsection{available filter functions}{
\itemize{
\item "sum" - sum of all cells in a MovingWindow
\item "min" - minimum value of all cells in a MovingWindow
\item "max" - maximum value of all cells in a MovingWindow
\item "mean"- mean value of all cells in a MovingWindow
\item "sd"  - standard deviation of all cells in a MovingWindow
\item "modal" - most frequent value of all cells in a MovingWindow
\item "sobel" - sobel edge detection filter in horizontal and vertical directions
\item "sobel_hrzt" - sobel edge detection filter in horizontal direction only
\item "sobel_vert" - sobel edge detection filter in vertical direction only
}
}
}
\examples{
### load data
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)
### compute all filters or every layer in stack
x <- filter_Stk(mspec,sizes=3)
x # note that the names are basic set x.1
### define layernames
ln <- c("blue","green","red","nir")
z <- filter_Stk(mspec,fLS="sum",sizes=3,layernames=ln)
names(z)
### compute specific filters
flist <- c("modal","sobel_vert","mean")
y <- filter_Stk(mspec,fLS=flist,sizes=c(3,5,7))
y
}
\seealso{
\code{\link{focal}},\code{\link{filter_Rst}}
}
\author{
Andreas Schönberg
}
