% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegInd_Nir.R
\name{vegInd_mspec}
\alias{vegInd_mspec}
\title{calculate RGB+Nir vegetation indices}
\usage{
vegInd_mspec(
  mspec,
  red = NULL,
  green = NULL,
  blue = NULL,
  nir = NULL,
  indlist = "all"
)
}
\arguments{
\item{mspec}{a RasterStack or Brick with multispectral RGB+Nir bands}

\item{red}{the band/layer number of band'red'}

\item{green}{the band/layer number of band'green'}

\item{blue}{the band/layer number of band'blue'}

\item{nir}{the band/layer number of band'Near Infrared (nir)'}

\item{indlist}{comma separated character for desired Vegetation Indices to compute. Select from
"NDVI","TDVI","SR","MSR" default=all (see details for further information)}
}
\value{
Returns a raster stack with the selected Vegetation Indices
}
\description{
computes several Vegetation Indices based on RGB bands
}
\details{
\subsection{available indices}{

"NDVI"
}
}
\note{
notes if needed
}
\examples{
### load data
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)
names(mspec)<- c("blue","green","red","nir")
### compute all vegetation indizes
x <-LEGION::vegInd_mspec(mspec,3,2,1,4)
plot(x)
### select specific vegetation indices
vi <-c("NDVI","TDVI")
y <-LEGION::vegInd_mspec(mspec,3,2,1,4,indlist=vi)
plot(y)
}
\references{
The IDB Project (2020): Index Database (https://www.indexdatabase.de/)
}
\author{
Andreas Schönberg
}
