% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegInd_RGB.R
\name{vegInd_RGB}
\alias{vegInd_RGB}
\title{Calculate RGB indices}
\usage{
vegInd_RGB(rgb, red = NULL, green = NULL, blue = NULL, indlist = "all")
}
\arguments{
\item{rgb}{a RasterStack with RGB bands}

\item{red}{the band/layer number of band 'red'}

\item{green}{the band/layer number of band 'green'}

\item{blue}{the band/layer number of band 'blue'}

\item{indlist}{comma-separated character combinations of the desired indices. Select from
"VVI","VARI","NDTI","RI","CI","BI","SI","HI","TGI","GLI","NGRDI", default=all. See details}
}
\value{
returns a RasterStack with the selected indices
}
\description{
computes several indices based on RGB bands
}
\details{
\subsection{available indices}{

"VVI" Visible Vegetation Index (1 - abs((red - 30) / (red + 30))) * (1 - abs((green - 50) / (green + 50))) \emph{(1 - abs((blue - 1) / (blue + 1)))
"VARI" Visible Atmospherically Resistant Index (green-red)/(green+red-blue)
"NDTI" Normalized Difference Turbidity Index (red-green)/(red+green)
"RI" Redness index (red**2/(blue}green\strong{3)
"CI" Soil Colour Index (red-green)/(red+green)
"BI" Brightness Index sqrt((red}2+green**2+blue\emph{2)/3)
"SI" Spectra Slope Saturation Index (red-blue)/(red+blue)
"HI" Primary Colours Hue Index (2}red-green-blue)/(green-blue)
"TGI" Triangular Greenness Index (-0.5*(190*(red - green)- 120*(red - blue))
"GLI" Green Leaf Index (2\emph{green-red-blue)/(2}green+red+blue)
"NGRDI" Normalized Green Red Difference Index (green-red)/(green+red)
}
}
\examples{
### load data
extpath <-system.file("extdata","lau_mspec.tif",package = "LEGION")
mspec <- raster::stack(extpath)
names(mspec)<- c("blue","green","red","nir")
### compute all vegetation indizes
x <-LEGION::vegInd_RGB(mspec,3,2,1)
plot(x)
### select specific vegetation indices
vi <-c("VVI","SI","GLI")
y <-LEGION::vegInd_RGB(mspec,3,2,1,indlist=vi)
plot(y)
}
\references{
The IDB Project (2020): Index Database (https://www.indexdatabase.de/)
}
\author{
Andreas Schönberg
}
