% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegInd_RGB.R
\name{vegInd_RGB}
\alias{vegInd_RGB}
\title{vegInd_RGB}
\usage{
vegInd_RGB(rgb, indlist = "all")
}
\arguments{
\item{rgb}{a RasterStack or Brick with RGB bands}

\item{indlist}{comma seperated character for desired Vegetation Indices to compute. Select from
"VVI","VARI","NDTI","RI","CI","BI","SI","HI","TGI","GLI","NGRDI", default=all}
}
\value{
Returns a raster stack with the selected Vegetation Indices
}
\description{
computes several Vegetation Indices based on RGB bands
}
\details{
further description
}
\note{
notes if needed
}
\examples{
# need to lern and test this
library(GSIF)
data(cookfarm)
### Prepare for 10-fold Leave-Location-and-Time-Out cross validation
indices <- CreateSpacetimeFolds(cookfarm$readings,"SOURCEID","Date")
str(indices)
### Prepare for 10-fold Leave-Location-Out cross validation
indices <- CreateSpacetimeFolds(cookfarm$readings,spacevar="SOURCEID")
str(indices)
### Prepare for leave-One-Location-Out cross validation
indices <- CreateSpacetimeFolds(cookfarm$readings,spacevar="SOURCEID",
k=length(unique(cookfarm$readings$SOURCEID)))
str(indices)
}
\references{
Meyer, H., Reudenbach, C., Hengl, T., Katurji, M., Nauß, T. (2018): Improving performance of spatio-temporal machine learning models using forward feature selection and target-oriented validation. Environmental Modelling & Software 101: 1-9.
}
\seealso{
\code{\link{trainControl}},\code{\link{ffs}} ### link to other functions
}
\author{
Andreas Schönberg
}
