% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegInd_RGB.R
\name{vegInd_RGB}
\alias{vegInd_RGB}
\title{vegInd_RGB}
\usage{
vegInd_RGB(rgb, indlist = "all")
}
\arguments{
\item{rgb}{a RasterStack or Brick with RGB bands}

\item{indlist}{comma separated character for desired Vegetation Indices to compute. Select from
"VVI","VARI","NDTI","RI","CI","BI","SI","HI","TGI","GLI","NGRDI", default=all (see details for further informations)}
}
\value{
Returns a raster stack with the selected Vegetation Indices
}
\description{
computes several Vegetation Indices based on RGB bands
}
\details{
further description
}
\note{
notes if needed
}
\examples{
# need to lern and test this
### load data
data(exp_rgb)
### compute all vegetation indizes
x <-LEGION::vegInd(exp_rgb)
plot(x)
### select specific vegetation indices
vi <-c("VVI","SI","GLI")
y <-LEGION::vegInd(exp_rgb,indlist=vi)
plot(y)
}
\references{
Meyer, H., Reudenbach, C., Hengl, T., Katurji, M., Nauß, T. (2018): Improving performance of spatio-temporal machine learning models using forward feature selection and target-oriented validation. Environmental Modelling & Software 101: 1-9.
}
\seealso{
\code{\link{trainControl}},\code{\link{ffs}} ### link to other functions
}
\author{
Andreas Schönberg
}
