% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegInd_RGB.R
\name{vegInd_RGB}
\alias{vegInd_RGB}
\title{calculate RGB vegetation indices}
\usage{
vegInd_RGB(rgb, indlist = "all")
}
\arguments{
\item{rgb}{a RasterStack or Brick with RGB bands}

\item{indlist}{comma separated character for desired Vegetation Indices to compute. Select from
"VVI","VARI","NDTI","RI","CI","BI","SI","HI","TGI","GLI","NGRDI", default=all (see details for further information)}
}
\value{
Returns a raster stack with the selected Vegetation Indices
}
\description{
computes several Vegetation Indices based on RGB bands
}
\details{
\subsection{available indices}{

"VVI"
"VARI"
"NDTI"
"RI"
"CI"
"BI"
"SI"
"HI"
"TGI"
"GLI"
"NGRDI"
}
}
\note{
notes if needed
}
\examples{
### load data
rgb_path <-system.file("extdata","exp_rgb.tif",package = "LEGION")
exp_rgb <- raster::stack(rgb_path)
### compute all vegetation indizes
x <-LEGION::vegInd_RGB(exp_rgb)
plot(x)
### select specific vegetation indices
vi <-c("VVI","SI","GLI")
y <-LEGION::vegInd(exp_rgb,indlist=vi)
plot(y)
}
\references{
The IDB Project (2020): Index Database (https://www.indexdatabase.de/)
}
\author{
Andreas Schönberg
}
